+++ app/build.gradle
@@ -37,7 +37,6 @@ dependencies {
     implementation project(path: ':basic:presentation')
     implementation project(path: ':basic:domain')
     implementation project(path: ':basic:data')
-    implementation project(path: ':paging')
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.0'
@@ -61,8 +60,13 @@ dependencies {
     // repeatOnLifeCycle
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.4.0"
 
-    implementation "com.google.dagger:hilt-android:2.40.5"
-    kapt "com.google.dagger:hilt-android-compiler:2.36"
+    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
+    testImplementation 'com.squareup.okhttp3:mockwebserver:4.9.1'
+    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+    implementation "com.squareup.okhttp3:logging-interceptor:4.9.1"
 
-    api("com.squareup.okhttp3:logging-interceptor:4.9.1")
+    implementation "com.google.dagger:hilt-android:2.40.5"
+    kapt 'com.google.dagger:hilt-compiler:2.40.5'
+
+    implementation 'androidx.activity:activity-ktx:1.4.0'
 }
