+++ app/src/main/java/com/survivalcoding/network_apps/feature_paging/domain/model/Post.kt
@@ -1,0 +1,8 @@
+package com.survivalcoding.network_apps.feature_paging.domain.model
+
+data class Post(
+    val userId: Int? = null,
+    val id: Int? = null,
+    val title: String? = null,
+    val body: String? = null
+)
