+++ app/src/main/java/com/survivalcoding/network_apps/feature_paging/domain/model/User.kt
@@ -1,0 +1,7 @@
+package com.survivalcoding.network_apps.feature_paging.domain.model
+
+data class User(
+    val id: Int? = null,
+    val name: String? = null,
+    val username: String? = null
+)
