+++ app/src/main/java/com/survivalcoding/network_apps/feature_paging/presentation/PostListActivity.kt
@@ -1,0 +1,22 @@
+package com.survivalcoding.network_apps.feature_paging.presentation
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.activity.viewModels
+import com.example.paging.databinding.ActivityPostListBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class PostListActivity : AppCompatActivity() {
+
+    private val viewModel: PostListViewModel by viewModels()
+
+    private val binding: ActivityPostListBinding by lazy {
+        ActivityPostListBinding.inflate(layoutInflater)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+    }
+}
