--- paging/src/main/java/com/example/paging/domain/model/Post.kt
@@ -1,8 +1,0 @@
-package com.example.paging.domain.model
-
-data class Post(
-    val userId: Int? = null,
-    val id: Int? = null,
-    val title: String? = null,
-    val body: String? = null
-)
