--- paging/src/main/java/com/example/paging/domain/model/User.kt
@@ -1,7 +1,0 @@
-package com.example.paging.domain.model
-
-data class User(
-    val id: Int? = null,
-    val name: String? = null,
-    val username: String? = null
-)
