--- paging/src/main/java/com/example/paging/domain/repository/JsonPlaceholderRepository.kt
@@ -1,9 +1,0 @@
-package com.example.paging.domain.repository
-
-import com.example.paging.domain.model.Post
-import com.example.paging.domain.model.User
-
-interface JsonPlaceholderRepository {
-    suspend fun getPosts(page: Int): List<Post>
-    suspend fun getUser(id: Int): User
-}
