--- paging/src/main/java/com/example/paging/presentation/PostListActivity.kt
@@ -1,12 +1,0 @@
-package com.example.paging.presentation
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.paging.R
-
-class PostListActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_post_list)
-    }
-}
