+++ settings.gradle
@@ -11,4 +11,3 @@ include ':app'
 include ':basic:presentation'
 include ':basic:data'
 include ':basic:domain'
-include ':paging'
